Gets bot playing a chime when you join.
When a user joins a voice channel in a guild the bot is in:

1. bot joins the voice channel
2. bot plays a chime

If another user joins the same channel while the bot is playing, it will
keep playing the original chime and overlay the new chime on top of it.

If another user joins a different channel, the bot will join the new
channel and play the chime regardless of whether or not the bot was
playing.