Adds docker-outside-of-docker support.
We need docker for the postgres database, which means we need to be able
to call docker from the devcontainer. The recommended way (since we
don't need to mount the file system) is to use the
`docker-outside-of-docker` approach.

This also updates the script for compatibility with the new approach. We
now need to connect to the host.docker.internal instead of localhost.

We can now start the database from within the dev container using:

POSTGRES_HOST=host.docker.internal ./scripts/init_db.sh