Adds URLs database to app.
This gets the app using the database to fetch the URL associated to the
user for that guild. For now these have to be manually added to the
database.

Requires migrating the database when building the app. We tell shuttle
to use the database in DATABASE_URL for local runs. Database gets stored
in the handler struct.

Gets test using the actual getter used by app.

Lots of improvements to make still but this gets it to a working state!