Set up autobahn websocket conformance tests
Big commit, these are the changes:

- moves websocket stuff to separate file
- implements websocket read loop
- hooks up autobahn conformance test suite
  - adds 'wstest.sh' for starting autobahn container
  - adds 'config/fuzzingserver.json' test configuration
  - gets the first test failing (gracefully, see 'reports' dir)
- removes Python websocket test server

Autobahn was a pain in the ass to figure out. The docs in the repo are
sparse and all the links are dead. Fortunately, the tungstenite-rs
project has a great, simple example of running the test suite. Thanks!

The test currently assumes the autobahn image is already running and
does not actually check whether or not tests pass (though I'm sure
there's a way to do that... close codes?). It does however generate a
nice report where you can currently find one failing test case.