Move control flow out of Client into application
The client no longer stays in a while loop calling a handler when
messages are available. It now reads until it has read a message and
returns the message to the application. The application is now
responsible for calling read in a loop.

We return an optional from read so that we can use the cool while loop
capture syntax. We return null when the connection should be closed.
That means we can just do a 'while (client.read(...) |msg| {...})' to
handle the connected state. After that point, the client is
disconnected: all future reads will return null.