Cleanup: address all TODOs and FIXMEs
Main change here is that we're no longer allocating on the heap during
the handshake. We instead reserve 512 bytes on the stack which should be
plenty to handle both the request and the response. This means the
client no longer requires an allocator.

While we were at it, modified how we were checking for end of HTTP
headers. Instead of reading one byte at a time we read one line at a
time until we get a line that is just "\r\n" (but without the "\n"
because it is used as a delimiter and the returned slice is not
inclusive of the delimiter). That is, we just have to check that the
first character in the line is "\r".

Also includes some other minor cleanups.